Convert User Groups link to abstract class for use with other post types
