+++ disciple-tools-bulk-magic-link-sender.php
@@ -90,6 +90,7 @@ class Disciple_Tools_Bulk_Magic_Link_Sender {
         require_once( 'magic-link/magic-links-api.php' );
         require_once( 'magic-link/magic-link-templates.php' );
         require_once( 'magic-link/magic-link-user-app.php' );
+        require_once( 'magic-link/magic-link-user-posts-base.php' );
         require_once( 'magic-link/magic-link-user-groups-app.php' );
         require_once( 'magic-link/magic-links-cron.php' );
         require_once( 'magic-link/magic-links-default-filters.php' );
